Switch to cmake + scikit-build for pybind11 extension compilation
Following the pybind11 scikit build example structure: https://github.com/pybind/scikit_build_example/
Switching to cmake + scikit-build offers a couple of advantages:
* Compilation at installation, instead at the first import (potentially blocking for a few minutes);
* Faster parallel compilation using Ninja
* Potential switch to Nanobind (cleaner & faster Python bindings);

The main drawback is additional build dependencies.

This PR is cleaning the Python and C++ organization of the project by:
* Switching to pyproject.toml config, removing setup.py (setup.cfg still need to be cleaned and removed);
* C++ code has been moved to a proper sub-directory `lib`;
* C++ compilation is split into 3 parts: core library, core Python bindings, and JAX bindings ops;
* cppimport dependency is removing;
