+++ .github/workflows/notebook-tests-public.yaml
@@ -31,16 +31,17 @@ jobs:
         id: pip-cache
         run: |
           python3 -m pip install --upgrade pip wheel setuptools jupyter
+          python3 -m pip install --upgrade build cmake ninja pybind11 scikit_build_core[pyproject]
       # Install JAX for IPU.
       - name: Install JAX for IPU
         run: |
           pip3 install --default-timeout=100 -U numpy==1.23.5 scipy etils pytest
           pip3 install jax==0.3.16+ipu jaxlib==0.3.15+ipu.sdk320 -f https://graphcore-research.github.io/jax-experimental/wheels.html
-      # Install repository.
+      # Build & install sdist package.
       - name: Install TessellateIPU
         run: |
-          python3 setup.py bdist_wheel --universal
-          pip3 install dist/*.whl
+          python3 -m build --no-isolation
+          pip3 install dist/*.tar.gz
       # Run TessellateIPU notebooks.
       - name: Run TessellateIPU notebooks (IPU model)
         run: |
