+++ .gitignore
@@ -17,9 +17,9 @@ dist/
 downloads/
 eggs/
 .eggs/
-lib/
 lib64/
 parts/
+reports/
 sdist/
 var/
 wheels/
