--- requirements.txt
@@ -1,6 +1,0 @@
-chex == 0.1.6
-cppimport
-jax >= 0.3.16
-jaxlib >= 0.3.15
-numpy >= 1.22.4
-pybind11
