+++ setup.cfg
@@ -1,9 +1,3 @@
-[tool:pytest]
-addopts = -p no:hypothesispytest
-markers =
-    ipu_hardware: Tests to run on IPU hardware.
-testpaths = tests
-
 [flake8]
 max-line-length = 120
 max-complexity = 20
