+++ tessellate_ipu/core/tile_common_utils.py
@@ -1,17 +1,10 @@
 # Copyright (c) 2022 Graphcore Ltd. All rights reserved.
-import os
 from typing import Any, Sequence
 
-import cppimport
 import numpy as np
 from numpy.typing import DTypeLike
 
-# Pybind11 extension import (and compilation if necessary).
-# Explicit path is more robust to different `pip install` usages.
-ext_filename = os.path.abspath(os.path.join(os.path.dirname(__file__), "tile_array_primitives_impl.cpp"))
-tile_array_primitives_impl = cppimport.imp_from_filepath(ext_filename, "tessellate_ipu.core.tile_array_primitives_impl")
-
-from .tile_array_primitives_impl import Base64Data, IpuShapedArray, IpuType  # noqa: E402, F401
+from tessellate_ipu.lib.pytessellate_ipu_core import Base64Data, IpuShapedArray, IpuType  # noqa: E402, F401
 
 _numpy_dtype_to_ipu_type = {
     np.dtype(np.bool_): IpuType.BOOL,
