+++ tessellate_ipu/lib/poplin/ConvPartialsStridesPacking.hpp
